Update RecursiveMenu.class.php
Since the constructor is the only place which needs the database it's imho better to just pass the arrays instead of the database. Also in my use-case I'm not using PDO but another db abstraction class. Instead of letting me change the constructor as soon as the database layer changes, let me pass arrays.